First and roughest version of disassembler
I know for a fact that the assembly output is going to be wildly wrong
since we don't know very much about how the data is organized, so treat
this command as experimental. We should be able to infer some
information with a bit of inspection -- that's my next step.